clean up to use fixed width ints

+++ sources/Application/Player/Player.cpp
@@ -874,7 +874,7 @@ void Player::playCursorPosition(int channel) {
 
         if (note < 128) {
           mixer_.StartInstrument(channel, instrument, note, newInstrument);
-          int instrTable = instrument->GetTable();
+          int8_t instrTable = instrument->GetTable();
 
           // If an instrument number has been specified && instrument has table,
           // we trigger the table.
@@ -953,7 +953,7 @@ void Player::StepAutomationTableForRetrigger(int channel,
 
 void Player::RetriggerChannelInstrument(int channel, int semitoneOffset,
                                         bool stepAutomationTable) {
-  int note = mixer_.GetChannelNote(channel);
+  int8_t note = mixer_.GetChannelNote(channel);
   I_Instrument *instrument = mixer_.GetInstrument(channel);
 
   if ((note > HIGHEST_NOTE) || (instrument == 0)) {
